+++ src/pages/blog.js
@@ -19,7 +19,7 @@ const Blog = ({ posts }) => {
     return (
         <div>
             <Nav />
-            <div className='flex-row max-w-4xl mt-24 mx-auto px-6 py-12 bg-gray-200 rounded-lg'>
+            <div className='flex flex-col max-w-4xl mt-24 mx-auto px-6 py-12 bg-gray-200 md:rounded-lg'>
                 <h3 className='ml-8 text-4xl text-teal-600 font-sans'>Posts</h3>
                 {posts ? (
                     sortPosts(posts).map((post) => (
